refactor(profile): modular structure and moved core stuff in utils
